Week 4/requirements.txt
sqlalchemy
prefect
requests
python-decouple
psycopg2
pandas
postgres
datetime

